Unassign vouchers and reset handler

+++ cashlesscards/deploy.py
@@ -60,7 +60,7 @@ def deploy_production():
     set_old_prod = "cashlesscards/settings_production.py"
     set_dev = "cashlesscards/settings_development.py"
     set_prod = "cashlesscards/settings.py"
-    if os.path.isfile(set_prod):
+    if os.path.isfile(set_prod) and os.path.isfile(set_old_prod):
         os.rename(set_prod, set_dev)
     if os.path.isfile(set_old_prod):
         os.rename(set_old_prod, set_prod)
